scripts/requirements.txt
generative-ai-hub-sdk[all]
flask
hdbcli
python-dotenv
flask_limiter
dbutils
PyPDF2
requests
pandas
pdfplumber
pypdf
numpy
opencv-python-headless
jpype1
camelot-py
tabula-py
pytesseract
pillow
openpyxl
fuzzywuzzy
reportlab
markdown
gunicorn==23.0.0
flask_cors
sap-xssec
